requirements.txt
streamlit
langchain==0.2.11
Langchain-openai